Update Quick Start Define Model code example
