+++ examples/kotlin/shared/src/commonMain/kotlin/com/mongodb/realm/realmkmmapp/RealmModels.kt
@@ -27,16 +27,3 @@ class Task : RealmObject {
     var name: String = "new task"
     var status: String = "Open"
 }
-
-// :snippet-start: quick-start-model
-class Item() : RealmObject {
-    @PrimaryKey
-    var _id: ObjectId = ObjectId()
-    var isComplete: Boolean = false
-    var summary: String = ""
-    var owner_id: String = ""
-    constructor(ownerId: String = "") : this() {
-        owner_id = ownerId
-    }
-}
-// :snippet-end:
