+++ source/sdk/kotlin/quick-start.txt
@@ -40,7 +40,7 @@ To define your application's data model, add a class definition to your
 application code. The example below illustrates the creation of an 'Item' model
 that represents Todo items in a Todo list app.
 
-.. literalinclude:: /examples/generated/kotlin/RealmModels.snippet.quick-start-model.kt
+.. literalinclude:: /examples/generated/kotlin/QuickStartTest.snippet.quick-start-model.kt
    :language: kotlin
    :copyable: false
